[Misc] Move CpuGpuBuffer to vllm/v1/utils.py (#23728)
